+++ src/main/kotlin/nl/deschepers/laraveltinker/action/ClearConsoleAction.kt
@@ -4,16 +4,15 @@ import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import nl.deschepers.laraveltinker.Strings
-import nl.deschepers.laraveltinker.repository.ConsoleFileRepository
+import nl.deschepers.laraveltinker.toolwindow.TinkerOutputToolWindowFactory
 import nl.deschepers.laraveltinker.util.getConsole
 
-/** Always opens a new tinker console */
-class RunConsoleAction : AnAction() {
+class ClearConsoleAction : AnAction() {
     override fun update(e: AnActionEvent) {
         super.update(e)
         e.presentation.isVisible = true
         e.presentation.isEnabled = false
-        e.presentation.text = Strings.get("lt.menu.action.run_console")
+        e.presentation.text = Strings.get("lt.menu.action.clear_console")
         e.project ?: return
         val console = e.getConsole()
         e.presentation.isEnabled = console != null
@@ -21,11 +20,10 @@ class RunConsoleAction : AnAction() {
 
     override fun actionPerformed(e: AnActionEvent) {
         val project = e.project ?: return
-        val consoleFileRepository = ConsoleFileRepository(project)
         val console = e.getConsole()
 
         if (console != null) {
-            consoleFileRepository.runFile(console)
+            TinkerOutputToolWindowFactory.tinkerOutputToolWindow[project]?.clearOutput()
         }
     }